Restructure files for new ES code
