+++ ixp_tracker/ixp_tracker.py
@@ -1,457 +1,17 @@
 import logging
-from dataclasses import dataclass
-from datetime import datetime, timezone, timedelta
-from enum import Enum
+from datetime import datetime, timedelta
 from typing import TypedDict, Any
 from uuid import uuid4
 
 from ixp_tracker.data_lookup import ASNGeoLookup
 from ixp_tracker.event_store import (
     EventStore,
-    Projection,
-    Aggregate,
-    DomainEvent,
     AggregateNotFound,
-    ValueNotChanged,
 )
-from ixp_tracker.models import StoredEvent, IXPIdMap, IXP as LegacyIXP, ASNMap
+import ixp_tracker.ixp_tracker_aggregates as ixpt
+from ixp_tracker.models import IXPIdMap, ASNMap
 
 logger = logging.getLogger("ixp_tracker")
-DATE_FORMAT = "%Y-%m-%d %H:%M:%S%z"
-
-
-@dataclass
-class IXPCreated(DomainEvent):
-    name: str
-    long_name: str
-    city: str
-    peeringdb_id: int
-    website: str
-    active_status: bool
-    country_code: str
-    created: str
-    last_updated: str
-    last_active: str
-    manrs_participant: bool
-    anchor_host: bool
-    org_id: int
-    physical_locations: int | None = None
-
-
-@dataclass
-class IXPUpdated(DomainEvent):
-    name: str | ValueNotChanged = ValueNotChanged()
-    long_name: str | ValueNotChanged = ValueNotChanged()
-    city: str | ValueNotChanged = ValueNotChanged()
-    website: str | ValueNotChanged = ValueNotChanged()
-    country_code: str | ValueNotChanged = ValueNotChanged()
-    created: str | ValueNotChanged = ValueNotChanged()
-    last_updated: str | ValueNotChanged = ValueNotChanged()
-    org_id: int | ValueNotChanged = ValueNotChanged()
-
-
-@dataclass
-class ManrsStatusChange(DomainEvent):
-    manrs_participant: bool
-
-
-@dataclass()
-class IXPBecameActive(DomainEvent):
-    active_status: bool
-
-    def __init__(self, active_status: bool = True):
-        if not active_status:
-            raise RuntimeError("IXPBecameActive must set status to True")
-        self.active_status = True
-
-
-@dataclass()
-class IXPBecameInactive(DomainEvent):
-    active_status: bool
-
-    def __init__(self, active_status: bool = False):
-        if active_status:
-            raise RuntimeError("IXPBecameInactive must set status to False")
-        self.active_status = False
-
-
-@dataclass
-class AnchorHostChange(DomainEvent):
-    anchor_host: bool
-
-
-@dataclass
-class PhysicalLocationChange(DomainEvent):
-    physical_locations: int
-
-
-@dataclass
-class IXPActiveInPeeringDb(DomainEvent):
-    last_active: str
-
-
-@dataclass
-class IXPMemberDetails:
-    date_joined: datetime
-    date_updated: datetime
-    last_active: datetime
-    is_rs_peer: bool
-    port_speed: int
-    date_left: datetime | None = None
-
-
-@dataclass
-class IXPMemberJoined(DomainEvent):
-    asn: int
-    date_joined: str
-    date_updated: str
-    last_active: str
-    is_rs_peer: bool
-    port_speed: int
-
-
-@dataclass
-class IXPMemberUpdated(DomainEvent):
-    asn: int
-    date_joined: str | ValueNotChanged = ValueNotChanged()
-    date_left: str | ValueNotChanged = ValueNotChanged()
-    date_updated: str | ValueNotChanged = ValueNotChanged()
-    port_speed: int | ValueNotChanged = ValueNotChanged()
-
-
-@dataclass
-class IXPMemberActiveInPeeringDb(DomainEvent):
-    asn: int
-    last_active: str
-
-
-@dataclass
-class RsPeeringStatusChange(DomainEvent):
-    asn: int
-    is_rs_peer: bool
-
-
-@dataclass
-class IXPMemberLeft(DomainEvent):
-    asn: int
-    date_left: str
-
-
-class IXP(Aggregate):
-    name: str
-    long_name: str
-    city: str
-    peeringdb_id: int
-    website: str
-    active_status: bool = True
-    country_code: str
-    date_created: datetime
-    last_updated: datetime
-    last_active: datetime
-    manrs_participant: bool = False
-    anchor_host: bool = False
-    org_id: int
-    physical_locations: int | None
-    members: dict[int, IXPMemberDetails]
-
-    def created(self, event: IXPCreated):
-        self.name = event.name
-        self.long_name = event.long_name
-        self.city = event.city
-        self.peeringdb_id = event.peeringdb_id
-        self.website = event.website
-        self.active_status = event.active_status
-        self.country_code = event.country_code
-        self.date_created = datetime.strptime(event.created, DATE_FORMAT)
-        self.last_updated = datetime.strptime(event.last_updated, DATE_FORMAT)
-        self.last_active = datetime.strptime(event.last_active, DATE_FORMAT)
-        self.manrs_participant = event.manrs_participant
-        self.anchor_host = event.anchor_host
-        self.org_id = event.org_id
-        self.physical_locations = event.physical_locations
-        self.members = {}
-
-    def updated(self, event: IXPUpdated):
-        if not isinstance(event.name, ValueNotChanged):
-            self.name = event.name
-        if not isinstance(event.long_name, ValueNotChanged):
-            self.long_name = event.long_name
-        if not isinstance(event.city, ValueNotChanged):
-            self.city = event.city
-        if not isinstance(event.website, ValueNotChanged):
-            self.website = event.website
-        if not isinstance(event.country_code, ValueNotChanged):
-            self.country_code = event.country_code
-        if not isinstance(event.created, ValueNotChanged):
-            self.date_created = datetime.strptime(event.created, DATE_FORMAT)
-        if not isinstance(event.last_updated, ValueNotChanged):
-            self.last_updated = datetime.strptime(event.last_updated, DATE_FORMAT)
-        if not isinstance(event.org_id, ValueNotChanged):
-            self.org_id = event.org_id
-
-    def became_active(self, event: IXPBecameActive):
-        self.active_status = event.active_status
-
-    def became_inactive(self, event: IXPBecameInactive):
-        self.active_status = event.active_status
-
-    def manrs_status_change(self, event: ManrsStatusChange):
-        self.manrs_participant = event.manrs_participant
-
-    def anchor_host_change(self, event: AnchorHostChange):
-        self.anchor_host = event.anchor_host
-
-    def physical_location_change(self, event: PhysicalLocationChange):
-        self.physical_locations = event.physical_locations
-
-    def active_in_peering_db(self, event: IXPActiveInPeeringDb):
-        self.last_active = datetime.strptime(event.last_active, DATE_FORMAT)
-
-    def get_members(
-        self, include_inactive: bool = False
-    ) -> dict[int, IXPMemberDetails]:
-        if include_inactive:
-            return self.members
-        member_list = {}
-        for member_asn in self.members.keys():
-            member = self.members[member_asn]
-            if member.date_left is None:
-                member_list[member_asn] = member
-        return member_list
-
-    def member_joined(self, event: IXPMemberJoined):
-        details = IXPMemberDetails(
-            datetime.strptime(event.date_joined, DATE_FORMAT),
-            datetime.strptime(event.date_updated, DATE_FORMAT),
-            datetime.strptime(event.last_active, DATE_FORMAT),
-            event.is_rs_peer,
-            event.port_speed,
-        )
-        self.members[event.asn] = details
-
-    def member_updated(self, event: IXPMemberUpdated):
-        if not isinstance(event.date_joined, ValueNotChanged):
-            self.members[event.asn].date_joined = datetime.strptime(
-                event.date_joined, DATE_FORMAT
-            )
-        if not isinstance(event.date_updated, ValueNotChanged):
-            self.members[event.asn].date_updated = datetime.strptime(
-                event.date_updated, DATE_FORMAT
-            )
-        if not isinstance(event.port_speed, ValueNotChanged):
-            self.members[event.asn].port_speed = event.port_speed
-        if not isinstance(event.date_left, ValueNotChanged):
-            self.members[event.asn].date_left = None
-
-    def member_active_in_peering_db(self, event: IXPMemberActiveInPeeringDb):
-        self.members[event.asn].last_active = datetime.strptime(
-            event.last_active, DATE_FORMAT
-        )
-
-    def rs_peering_status_change(self, event: RsPeeringStatusChange):
-        self.members[event.asn].is_rs_peer = event.is_rs_peer
-
-    def member_left(self, event: IXPMemberLeft):
-        self.members[event.asn].date_left = datetime.strptime(
-            event.date_left, DATE_FORMAT
-        )
-
-    def snapshot(self):
-        values = super().snapshot()
-        values["date_created"] = stringify_date(values["date_created"])
-        values["last_updated"] = stringify_date(values["last_updated"])
-        values["last_active"] = stringify_date(values["last_active"])
-        values["members"] = dict(values["members"])
-        for member_asn in values["members"].keys():
-            values["members"][member_asn] = dict(values["members"][member_asn].__dict__)
-            values["members"][member_asn]["date_joined"] = stringify_date(
-                values["members"][member_asn]["date_joined"]
-            )
-            values["members"][member_asn]["date_updated"] = stringify_date(
-                values["members"][member_asn]["date_updated"]
-            )
-            values["members"][member_asn]["last_active"] = stringify_date(
-                values["members"][member_asn]["last_active"]
-            )
-            if values["members"][member_asn]["date_left"]:
-                values["members"][member_asn]["date_left"] = stringify_date(
-                    values["members"][member_asn]["date_left"]
-                )
-        return values
-
-    def hydrate(self, data: dict):
-        super().hydrate(data)
-        self.date_created = datetime.strptime(data["date_created"], DATE_FORMAT)
-        self.last_updated = datetime.strptime(data["last_updated"], DATE_FORMAT)
-        self.last_active = datetime.strptime(data["last_active"], DATE_FORMAT)
-        members = {}
-        for member_asn in self.members.keys():
-            member_details = self.members[member_asn]
-            member_details["date_joined"] = datetime.strptime(  # type: ignore
-                self.members[member_asn]["date_joined"],  # type: ignore
-                DATE_FORMAT,
-            )
-            member_details["date_updated"] = datetime.strptime(  # type: ignore
-                self.members[member_asn]["date_updated"],  # type: ignore
-                DATE_FORMAT,
-            )
-            member_details["last_active"] = datetime.strptime(  # type: ignore
-                self.members[member_asn]["last_active"],  # type: ignore
-                DATE_FORMAT,
-            )
-            if member_details["date_left"]:  # type: ignore
-                member_details["date_left"] = datetime.strptime(  # type: ignore
-                    self.members[member_asn]["date_left"],  # type: ignore
-                    DATE_FORMAT,
-                )
-            members[int(member_asn)] = IXPMemberDetails(**member_details)  # type: ignore
-        self.members = members
-
-
-class NetworkType(Enum):
-    NSP = "NSP"
-    CONTENT = "Content"
-    CABLE_DSL_ISP = "Cable/DSL/ISP"
-    ENTERPRISE = "Enterprise"
-    EDUCATION_RESEARCH = "Educational/Research"
-    NON_PROFIT = "Non-Profit"
-    ROUTE_SERVER = "Route Server"
-    NETWORK_SERVICES = "Network Services"
-    ROUTE_COLLECTOR = "Route Collector"
-    GOVERNMENT = "Government"
-    NOT_DISCLOSED = "Not Disclosed"
-    OTHER = "Other"
-    UNKNOWN = "Unknown"
-
-
-class PeeringPolicy(Enum):
-    OPEN = "Open"
-    SELECTIVE = "Selective"
-    RESTRICTIVE = "Restrictive"
-    NO = "No"
-    UNKNOWN = "Unknown"
-
-
-@dataclass
-class ASNCreated(DomainEvent):
-    as_number: int
-    name: str
-    network_type: str
-    peering_policy: str
-    peeringdb_id: int
-    country_code: str
-
-
-@dataclass
-class ASNUpdated(DomainEvent):
-    name: str | ValueNotChanged = ValueNotChanged()
-    network_type: str | ValueNotChanged = ValueNotChanged()
-    peering_policy: str | ValueNotChanged = ValueNotChanged()
-    country_code: str | ValueNotChanged = ValueNotChanged()
-
-
-# We don't think this should ever happen but record as a separate event if it does
-@dataclass
-class ASNPeeringDbIdChanged(DomainEvent):
-    peeringdb_id: int
-
-
-IXP_TRACKER_EVENT_MAP = {
-    AnchorHostChange.__name__: AnchorHostChange,
-    ASNCreated.__name__: ASNCreated,
-    ASNPeeringDbIdChanged.__name__: ASNPeeringDbIdChanged,
-    ASNUpdated.__name__: ASNUpdated,
-    ManrsStatusChange.__name__: ManrsStatusChange,
-    IXPActiveInPeeringDb.__name__: IXPActiveInPeeringDb,
-    IXPBecameActive.__name__: IXPBecameActive,
-    IXPBecameInactive.__name__: IXPBecameInactive,
-    IXPCreated.__name__: IXPCreated,
-    IXPUpdated.__name__: IXPUpdated,
-    IXPMemberActiveInPeeringDb.__name__: IXPMemberActiveInPeeringDb,
-    IXPMemberJoined.__name__: IXPMemberJoined,
-    IXPMemberLeft.__name__: IXPMemberLeft,
-    IXPMemberUpdated.__name__: IXPMemberUpdated,
-    PhysicalLocationChange.__name__: PhysicalLocationChange,
-    RsPeeringStatusChange.__name__: RsPeeringStatusChange,
-}
-
-
-class ASN(Aggregate):
-    name: str
-    number: int
-    peeringdb_id: int
-    # NetworkType is being retired in a future version of PeeringDb so we may need to look for a different source for this
-    network_type: NetworkType
-    peering_policy: PeeringPolicy
-    country_code: str
-
-    def created(self, event: ASNCreated):
-        self.name = event.name
-        self.number = event.as_number
-        self.peeringdb_id = event.peeringdb_id
-        self.network_type = NetworkType(event.network_type)
-        self.peering_policy = PeeringPolicy(event.peering_policy)
-        self.country_code = event.country_code
-
-    def updated(self, event: ASNUpdated):
-        if not isinstance(event.name, ValueNotChanged):
-            self.name = event.name
-        if not isinstance(event.network_type, ValueNotChanged):
-            self.network_type = NetworkType(event.network_type)
-        if not isinstance(event.peering_policy, ValueNotChanged):
-            self.peering_policy = PeeringPolicy(event.peering_policy)
-        if not isinstance(event.country_code, ValueNotChanged):
-            self.country_code = event.country_code
-
-    def peering_db_id_changed(self, event: ASNPeeringDbIdChanged):
-        self.peeringdb_id = event.peeringdb_id
-
-
-class IXPIdMapProjection(Projection):
-    aggregate_types = [IXP.__name__]
-    events = [IXPCreated.__name__]
-
-    def do_handle(self, event: StoredEvent, aggregate: Aggregate):
-        existing = IXPIdMap.objects.filter(aggregate_id=event.aggregate_id)
-        if existing.count() > 0:
-            return
-        peeringdb_id = event.data.get("peeringdb_id", None)
-        preserve_legacy_id = LegacyIXP.objects.filter(peeringdb_id=peeringdb_id).first()
-        if preserve_legacy_id:
-            # If a legacy object exists, use that's primary key as our primary key to preserve the "isoc_id"
-            isoc_id = IXPIdMap(
-                id=preserve_legacy_id.pk,
-                aggregate_id=event.aggregate_id,
-                peeringdb_id=peeringdb_id,
-            )
-        else:
-            isoc_id = IXPIdMap(
-                aggregate_id=event.aggregate_id,
-                peeringdb_id=peeringdb_id,
-            )
-        isoc_id.save()
-
-    def find_by_peeringdb_id(self, peeringdb_id: int) -> IXPIdMap | None:
-        try:
-            return IXPIdMap.objects.get(peeringdb_id=peeringdb_id)
-        except IXPIdMap.DoesNotExist:
-            return None
-
-
-class ASNList(Projection):
-    aggregate_types = [ASN.__name__]
-    events = [ASNCreated.__name__]
-
-    def do_handle(self, event: StoredEvent, asn: Aggregate):
-        existing = ASNMap.objects.filter(aggregate_id=event.aggregate_id)
-        if existing.count() > 0:
-            return
-        asn = event.data.get("as_number", None)
-        asn_map = ASNMap(
-            aggregate_id=event.aggregate_id,
-            asn=asn,
-        )
-        asn_map.save()
 
 
 class MemberImportData(TypedDict):
@@ -501,8 +61,8 @@ class IXPTracker:
                 physical_locations,
             )
         active_status = False
-        ixp = IXP(id=uuid4())
-        event = IXPCreated(
+        ixp = ixpt.IXP(id=uuid4())
+        event = ixpt.IXPCreated(
             name,
             long_name,
             city,
@@ -510,9 +70,9 @@ class IXPTracker:
             website,
             active_status,
             country_code,
-            stringify_date(created),
-            stringify_date(last_updated),
-            stringify_date(last_active),
+            ixpt.stringify_date(created),
+            ixpt.stringify_date(last_updated),
+            ixpt.stringify_date(last_active),
             manrs_participant,
             anchor_host,
             org_id,
@@ -523,7 +83,7 @@ class IXPTracker:
 
     def _update_ixp(
         self,
-        ixp: IXP,
+        ixp: ixpt.IXP,
         name: str,
         long_name: str,
         city: str,
@@ -549,29 +109,31 @@ class IXPTracker:
         if country_code != ixp.country_code:
             updates["country_code"] = country_code
         if created != ixp.date_created:
-            updates["created"] = stringify_date(created)
+            updates["created"] = ixpt.stringify_date(created)
         if last_updated != ixp.last_updated:
-            updates["last_updated"] = stringify_date(last_updated)
+            updates["last_updated"] = ixpt.stringify_date(last_updated)
         if org_id != ixp.org_id:
             updates["org_id"] = org_id
         if len(updates.keys()) > 0:
-            event = IXPUpdated(**updates)
+            event = ixpt.IXPUpdated(**updates)
             ixp = self.es.store(ixp, event)
         if ixp.manrs_participant != manrs_participant:
-            manrs_update = ManrsStatusChange(manrs_participant=manrs_participant)
+            manrs_update = ixpt.ManrsStatusChange(manrs_participant=manrs_participant)
             ixp = self.es.store(ixp, manrs_update)
         if ixp.anchor_host != anchor_host:
-            anchor_host_event = AnchorHostChange(anchor_host=anchor_host)
+            anchor_host_event = ixpt.AnchorHostChange(anchor_host=anchor_host)
             ixp = self.es.store(ixp, anchor_host_event)
         if (
             ixp.physical_locations != physical_locations
             and physical_locations is not None
         ):
-            locations_event = PhysicalLocationChange(
+            locations_event = ixpt.PhysicalLocationChange(
                 physical_locations=physical_locations
             )
             ixp = self.es.store(ixp, locations_event)
-        active_event = IXPActiveInPeeringDb(last_active=stringify_date(last_active))
+        active_event = ixpt.IXPActiveInPeeringDb(
+            last_active=ixpt.stringify_date(last_active)
+        )
         ixp = self.es.store(ixp, active_event)
         return ixp
 
@@ -579,8 +141,8 @@ class IXPTracker:
         self,
         as_number: int,
         name: str,
-        network_type: NetworkType,
-        peering_policy: PeeringPolicy,
+        network_type: ixpt.NetworkType,
+        peering_policy: ixpt.PeeringPolicy,
         peeringdb_id: int,
         country_code,
     ):
@@ -596,14 +158,14 @@ class IXPTracker:
             if country_code != entity.country_code:
                 updates["country_code"] = country_code
             if len(updates.keys()) > 0:
-                update_event = ASNUpdated(**updates)
+                update_event = ixpt.ASNUpdated(**updates)
                 entity = self.es.store(entity, update_event)
             if peeringdb_id != entity.peeringdb_id:
-                peering_id_event = ASNPeeringDbIdChanged(peeringdb_id=peeringdb_id)
+                peering_id_event = ixpt.ASNPeeringDbIdChanged(peeringdb_id=peeringdb_id)
                 entity = self.es.store(entity, peering_id_event)
         else:
-            entity = ASN(id=uuid4())
-            event = ASNCreated(
+            entity = ixpt.ASN(id=uuid4())
+            event = ixpt.ASNCreated(
                 as_number,
                 name,
                 network_type.value,
@@ -616,10 +178,10 @@ class IXPTracker:
 
     def import_members(
         self,
-        ixp: IXP,
+        ixp: ixpt.IXP,
         ixp_data: list[MemberImportData],
         processing_date: datetime,
-    ) -> IXP:
+    ) -> ixpt.IXP:
         existing_members = ixp.get_members(True)
         for member in ixp_data:
             as_entity = self.get_asn(member["asn"])
@@ -633,11 +195,11 @@ class IXPTracker:
                 and existing_member.date_left < member["created_date"]
             )
             if existing_member is None or member_has_rejoined:
-                join_event = IXPMemberJoined(
+                join_event = ixpt.IXPMemberJoined(
                     member["asn"],
-                    stringify_date(member["created_date"]),
-                    stringify_date(member["updated_date"]),
-                    stringify_date(processing_date),
+                    ixpt.stringify_date(member["created_date"]),
+                    ixpt.stringify_date(member["updated_date"]),
+                    ixpt.stringify_date(processing_date),
                     member["is_rs_peer"],
                     member["port_speed"],
                 )
@@ -652,21 +214,23 @@ class IXPTracker:
                 ):
                     updates["date_left"] = None
                 elif member["created_date"] != existing_member.date_joined:
-                    updates["date_joined"] = stringify_date(member["created_date"])
+                    updates["date_joined"] = ixpt.stringify_date(member["created_date"])
                 if member["updated_date"] != existing_member.date_updated:
-                    updates["date_updated"] = stringify_date(member["updated_date"])
+                    updates["date_updated"] = ixpt.stringify_date(
+                        member["updated_date"]
+                    )
                 if member["port_speed"] != existing_member.port_speed:
                     updates["port_speed"] = member["port_speed"]
                 if len(updates.keys()) > 0:
-                    update_event = IXPMemberUpdated(member["asn"], **updates)
+                    update_event = ixpt.IXPMemberUpdated(member["asn"], **updates)
                     ixp = self.es.store(ixp, update_event)
                 if member["is_rs_peer"] != existing_member.is_rs_peer:
-                    rs_peer_event = RsPeeringStatusChange(
+                    rs_peer_event = ixpt.RsPeeringStatusChange(
                         member["asn"], member["is_rs_peer"]
                     )
                     ixp = self.es.store(ixp, rs_peer_event)
-                active_event = IXPMemberActiveInPeeringDb(
-                    member["asn"], stringify_date(processing_date)
+                active_event = ixpt.IXPMemberActiveInPeeringDb(
+                    member["asn"], ixpt.stringify_date(processing_date)
                 )
                 ixp = self.es.store(ixp, active_event)
 
@@ -674,52 +238,48 @@ class IXPTracker:
         self.es.save_snapshot(ixp)
         return ixp
 
-    def check_ixp_inactive(self, ixp: IXP, processing_date: datetime) -> IXP:
+    def check_ixp_inactive(self, ixp: ixpt.IXP, processing_date: datetime) -> ixpt.IXP:
         members = ixp.get_members()
         members_left = check_if_members_have_left(
             members, processing_date, self.geo_lookup
         )
         for member_left in members_left:
-            left_event = IXPMemberLeft(member_left[0], stringify_date(member_left[1]))
+            left_event = ixpt.IXPMemberLeft(
+                member_left[0], ixpt.stringify_date(member_left[1])
+            )
             ixp = self.es.store(ixp, left_event)
         members = ixp.get_members()
         if ixp.active_status is False and len(members) >= 3:
-            active_event = IXPBecameActive()
+            active_event = ixpt.IXPBecameActive()
             ixp = self.es.store(ixp, active_event)
         elif ixp.active_status is True and len(members) < 3:
-            inactive_event = IXPBecameInactive()
+            inactive_event = ixpt.IXPBecameInactive()
             ixp = self.es.store(ixp, inactive_event)
         return ixp
 
-    def find_by_peeringdb_id(self, peeringdb_id: int) -> IXP | None:
+    def find_by_peeringdb_id(self, peeringdb_id: int) -> ixpt.IXP | None:
         try:
             id_map = IXPIdMap.objects.get(peeringdb_id=peeringdb_id)
-            return self.es.get_aggregate(id_map.aggregate_id, IXP)
+            return self.es.get_aggregate(id_map.aggregate_id, ixpt.IXP)
         except (IXPIdMap.DoesNotExist, AggregateNotFound):
             return None
 
     def get_all_ixps(self):
-        return self.es.get_all(IXP)
+        return self.es.get_all(ixpt.IXP)
 
     def get_all_asns(self):
-        return self.es.get_all(ASN)
+        return self.es.get_all(ixpt.ASN)
 
-    def get_asn(self, asn) -> ASN | None:
+    def get_asn(self, asn) -> ixpt.ASN | None:
         try:
             asn_map = ASNMap.objects.get(asn=asn)
-            return self.es.get_aggregate(asn_map.aggregate_id, ASN)
+            return self.es.get_aggregate(asn_map.aggregate_id, ixpt.ASN)
         except ASNMap.DoesNotExist:
             return None
 
 
-def stringify_date(date_value: datetime) -> str:
-    if date_value.tzinfo is None:
-        date_value = date_value.replace(tzinfo=timezone.utc)
-    return date_value.strftime(DATE_FORMAT)
-
-
 def check_if_members_have_left(
-    members: dict[int, IXPMemberDetails],
+    members: dict[int, ixpt.IXPMemberDetails],
     processing_date: datetime,
     geo_lookup: ASNGeoLookup,
 ) -> list[tuple[int, datetime]]:
