+++ tests/fixtures.py
@@ -21,20 +21,21 @@ from ixp_tracker.event_store import (
 )
 from ixp_tracker.ixp_tracker import (
     IXPTracker,
-    IXP_TRACKER_EVENT_MAP,
-    ASNList,
-    IXPIdMapProjection,
-    IXP,
+)
+from ixp_tracker.ixp_tracker_aggregates import (
     IXPCreated,
-    ASN,
-    NetworkType,
-    PeeringPolicy,
-    ASNCreated,
-    stringify_date,
+    IXPBecameActive,
     IXPMemberJoined,
     IXPMemberLeft,
-    IXPBecameActive,
+    ASNCreated,
+    IXP_TRACKER_EVENT_MAP,
+    ASN,
+    IXP,
+    NetworkType,
+    PeeringPolicy,
+    stringify_date,
 )
+from ixp_tracker.ixp_tracker_projections import ASNList, IXPIdMapProjection
 import ixp_tracker.models as legacy
 from ixp_tracker.models import (
     StatsPerCountry,
