+++ tests/test_members_import_es.py
@@ -6,11 +6,9 @@ from ixp_tracker import importers
 from ixp_tracker.event_store import EventStorePersistence, EventStore, DjangoEventStore
 from ixp_tracker.ixp_tracker import (
     IXPTracker,
-    IXP_TRACKER_EVENT_MAP,
-    IXPIdMapProjection,
-    ASNList,
-    IXP,
 )
+from ixp_tracker.ixp_tracker_aggregates import IXP_TRACKER_EVENT_MAP, IXP
+from ixp_tracker.ixp_tracker_projections import ASNList, IXPIdMapProjection
 from tests.fixtures import (
     PeeringNetIXLANFactory,
     TestLookup,
