+++ tests/test_snapshot_ixps.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timezone
 
 from ixp_tracker.event_store import EventStore
-from ixp_tracker.ixp_tracker import IXP_TRACKER_EVENT_MAP, IXP
+from ixp_tracker.ixp_tracker_aggregates import IXP_TRACKER_EVENT_MAP, IXP
 from tests.fixtures import create_ixp, create_member, create_asn, MemoryEventStore
